Remove all references to nlohmann and replace with ArduinoJson

+++ src/modules/control_tasks/StageControl.cpp
@@ -54,7 +54,7 @@ double StageControl::calculate_status() const {
     } else if (current_stage == Stage::PRESSURIZATION) {
         // if we're using PT-2 in the current test
         if(global_registry.sensors["pressure"].find("PT-2") != global_registry.sensors["pressure"].end()) {
-            double pressure = global_registry.sensors["pressure"]["PT-2"].normalized_value;
+            float pressure = global_registry.sensors["pressure"]["PT-2"].normalized_value;
             return std::min(100.0, pressure/4.9);
         }
         else {
